+++ xxl-job-admin/src/main/java/com/xxl/job/admin/core/thread/JobCompleteHelper.java
@@ -78,10 +78,10 @@ public class JobCompleteHelper {
                     try {
                         // 任务结果丢失处理：调度记录停留在 "运行中" 状态超过10min，且对应执行器心跳注册失败不在线，则将本地调度主动标记失败；
                         Date losedTime = DateUtil.addMinutes(new Date(), -10);
-                        List<Long> losedJobIds = XxlJobAdminConfig.getAdminConfig().getXxlJobLogDao().findLostJobIds(losedTime);
+                        List<Integer> losedJobIds = XxlJobAdminConfig.getAdminConfig().getXxlJobLogDao().findLostJobIds(losedTime);
 
                         if (losedJobIds != null && losedJobIds.size() > 0) {
-                            for (Long logId : losedJobIds) {
+                            for (Integer logId : losedJobIds) {
 
                                 XxlJobLog jobLog = new XxlJobLog();
                                 jobLog.setId(logId);
